ADD test cases to test the functionality and ADD .gitignore to ignore the unnecessary files and Change the program structure by breaking it to different files
