+++ DC-322-L-1 Web Engineering Lab/atmData.txt
@@ -1,6 +1,2 @@
-ATM001,2312,55010
-ATM098,1234,400
-ATM111,1111,100
-ATM555,5555,888
-ATM444,3434,100
-ATM098,1111,100
+ATM001,1234,500
+ATM098,5678,1000
